[tunip-1] fix: support spark on aws
